+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 - Fixed sea creature glow only displaying one entity and invisible entities sometimes
 - Fixed bossbar and merge health settings in /rfuscedit not properly applying to sea creatures
+- Peek party finder keybind now closes the creation window / presets
 
 ### Back-end
 - Reworked entity system to accept more model entities